Moved module processing to module.cpp
